:rocket: feat: improve load times 5x
