+++ packages/client/src/pages/officer/index.tsx
@@ -7,17 +7,7 @@ import { getSessionUser } from "lib/auth";
 import { getTranslations } from "lib/getTranslation";
 import type { GetServerSideProps } from "next";
 import { ActiveOfficer, useLeoState } from "state/leoState";
-import {
-  Bolo,
-  CombinedLeoUnit,
-  EmsFdDeputy,
-  LeoIncident,
-  Officer,
-  Record,
-  RecordType,
-  ShouldDoType,
-  ValueType,
-} from "@snailycad/types";
+import { Bolo, EmsFdDeputy, LeoIncident, Record, RecordType, ValueType } from "@snailycad/types";
 import { ActiveCalls } from "components/leo/ActiveCalls";
 import { Full911Call, useDispatchState } from "state/dispatchState";
 import { ModalButtons } from "components/leo/ModalButtons";
@@ -33,7 +23,6 @@ import { useModal } from "state/modalState";
 import { ModalIds } from "types/ModalIds";
 import { Permissions } from "@snailycad/permissions";
 import { useNameSearch } from "state/search/nameSearchState";
-import { useAuth } from "context/AuthContext";
 import { useFeatureEnabled } from "hooks/useFeatureEnabled";
 import { useTones } from "hooks/global/useTones";
 import { useLoadValuesClientSide } from "hooks/useLoadValuesClientSide";
@@ -76,8 +65,8 @@ interface Props {
   bolos: Bolo[];
   activeIncidents: LeoIncident[];
 
-  allDeputies: EmsFdDeputy[];
-  allOfficers: Officer[];
+  activeDeputies: EmsFdDeputy[];
+  activeOfficers: ActiveOfficer[];
 }
 
 export default function OfficerDashboard({
@@ -86,8 +75,8 @@ export default function OfficerDashboard({
   activeOfficer,
   activeIncidents,
 
-  allOfficers,
-  allDeputies,
+  activeOfficers,
+  activeDeputies,
 }: Props) {
   useLoadValuesClientSide({
     valueTypes: [
@@ -110,7 +99,6 @@ export default function OfficerDashboard({
   const tones = useTones("leo");
   const panic = usePanicButton();
   const { isOpen } = useModal();
-  const { user } = useAuth();
   const { LEO_TICKETS } = useFeatureEnabled();
 
   const { currentResult, setCurrentResult } = useNameSearch();
@@ -131,28 +119,12 @@ export default function OfficerDashboard({
     dispatchState.setBolos(bolos);
 
     dispatchState.setActiveIncidents(activeIncidents);
-    dispatchState.setAllOfficers(allOfficers);
-
-    function activeFilter(v: EmsFdDeputy | Officer | CombinedLeoUnit) {
-      return Boolean(v.statusId && v.status?.shouldDo !== ShouldDoType.SET_OFF_DUTY);
-    }
-
-    const activeOfficers = [...allOfficers].filter(activeFilter);
-    const activeDeputies = [...allDeputies].filter(activeFilter);
 
     dispatchState.setActiveDeputies(activeDeputies);
     dispatchState.setActiveOfficers(activeOfficers);
 
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [bolos, calls, allOfficers, allDeputies, activeOfficer]);
-
-  React.useEffect(() => {
-    if (!user) return;
-    const userOfficers = allOfficers.filter((v) => v.userId === user.id);
-
-    leoState.setOfficers(userOfficers);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [user, allOfficers]);
+  }, [bolos, calls, activeOfficers, activeDeputies, activeOfficer]);
 
   return (
     <Layout
@@ -216,31 +188,30 @@ export default function OfficerDashboard({
   );
 }
 
-export const getServerSideProps: GetServerSideProps = async ({ req, locale }) => {
-  const [
-    activeOfficer,
-    values,
-    calls,
-    bolos,
-    { officers: allOfficers, deputies: allDeputies, activeIncidents },
-  ] = await requestAll(req, [
-    ["/leo/active-officer", null],
-    ["/admin/values/codes_10", []],
-    ["/911-calls", []],
-    ["/bolos", []],
-    ["/dispatch", { officers: [], deputies: [], activeIncidents: [] }],
-  ]);
+export const getServerSideProps: GetServerSideProps<Props> = async ({ req, locale }) => {
+  const [activeOfficer, values, calls, bolos, activeOfficers, activeDeputies] = await requestAll(
+    req,
+    [
+      ["/leo/active-officer", null],
+      ["/admin/values/codes_10", []],
+      ["/911-calls", []],
+      ["/bolos", []],
+      ["/leo/active-officers", []],
+      ["/ems-fd/active-deputies", []],
+    ],
+  );
 
   return {
     props: {
+      // todo:
+      activeIncidents: [],
       session: await getSessionUser(req),
-      allOfficers,
-      allDeputies,
+      activeOfficers,
+      activeDeputies,
       activeOfficer,
       calls,
       bolos,
       values,
-      activeIncidents,
       messages: {
         ...(await getTranslations(
           ["citizen", "leo", "truck-logs", "ems-fd", "calls", "common"],
